LayoutTests/fast/forms/incremental-dom-property.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test for <a href="https://bugs.webkit.org/show_bug.cgi?id=50335">https://bugs.webkit.org/show_bug.cgi?id=50335</a>');

var input = document.createElement('input');

input.setAttribute('incremental', '')
shouldBeTrue('input.incremental');

input.removeAttribute('incremental');
shouldBeFalse('input.incremental');

input.incremental = true;
shouldBeTrue('input.hasAttribute("incremental")');

input.incremental = false;
shouldBeFalse('input.hasAttribute("incremental")');
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
